Delete button now functional, search feature now works for type1
delete button now access id without issue, and cleaned up html file so delete button now functions as intended. Navbar search button now changes the value looked for in the query allowing for filtering based off type1

+++ views.py
@@ -49,7 +49,9 @@ def init_routes(app):
     def delete_item(id):
         # This route should handle deleting an existing item identified by the given ID.
         item = Pokemon.query.get(id)  # Fetch item by ID
-
+        if item is None:
+            flash("Item not found", "error")
+            return redirect(url_for("index"))
         db.session.delete(item)  	# Delete item
 
         db.session.commit()  		# Commit changes
